+++ operators/src/pro/cache/mod.rs
@@ -1,8 +1,6 @@
-pub mod cache_chunk_stream;
 pub mod cache_chunks;
 pub mod cache_operator;
-pub mod cache_tile_stream;
 pub mod cache_tiles;
 pub mod error;
-pub mod tile_cache;
+pub mod shared_cache;
 pub mod util;
